refactor: cleaned up config
