+++ TrackingBackend/app/config/etvr_config.py
@@ -1,93 +1,12 @@
 from __future__ import annotations
 import json
 import os.path
-import re
-from pydantic import BaseModel, ValidationError, field_validator
-from .logger import get_logger
 from fastapi import Request, HTTPException
-from app.types import Algorithms
+from pydantic import BaseModel, ValidationError
+from app.config import AlgorithmConfig, CameraConfig, OSCConfig, CONFIG_FILE
+from app.logger import get_logger
 
 logger = get_logger()
-
-CONFIG_FILE = "tracker-config.json"
-# https://regex101.com/r/qlLITU/1
-IP_ADDRESS_REGEX = (
-    r"(\b(?:http:\/\/)?(?:(?:25[0-5]|2[0-4]\d|[01]?\d\d?)\.){3}(?:25[0-5]|2[0-4]\d|[01]?\d\d?)"
-    r"(?::\d{1,5})?\b|localhost(?::\d{1,5})?|http:\/\/localhost(?::\d{1,5})?|[\w-]+\.local(?::\d{1,5})?)"
-)
-
-
-class BlobConfig(BaseModel):
-    threshold: int = 65
-    minsize: int = 10
-    maxsize: int = 25
-
-
-class AlgorithmConfig(BaseModel):
-    algorithm_order: list[Algorithms] = [Algorithms.BLOB, Algorithms.HSRAC, Algorithms.RANSAC, Algorithms.HSF]
-    blob: BlobConfig = BlobConfig()
-
-
-class OSCConfigEndpoints(BaseModel):
-    eyes_y: str = "/avatar/parameters/EyesY"
-    left_eye_x: str = "/avatar/parameters/LeftEyeX"
-    right_eye_x: str = "/avatar/parameters/RightEyeX"
-    recenter: str = "/avatar/parameters/etvr_recenter"
-    sync_blink: str = "/avatar/parameters/etvr_sync_blink"
-    recalibrate: str = "/avatar/parameters/etvr_recalibrate"
-    left_eye_blink: str = "/avatar/parameters/LeftEyeLidExpandedSqueeze"
-    right_eye_blink: str = "/avatar/parameters/RightEyeLidExpandedSqueeze"
-
-
-class OSCConfig(BaseModel):
-    address: str = "127.0.0.1"
-    mirror_eyes: bool = False
-    sync_blink: bool = False
-    enable_sending: bool = True
-    sending_port: int = 9000
-    enable_receiving: bool = False
-    receiver_port: int = 9001
-    vrchat_native_tracking: bool = False
-    endpoints: OSCConfigEndpoints = OSCConfigEndpoints()
-
-    @field_validator("address")
-    def address_validator(cls, value: str) -> str:
-        if re.match(IP_ADDRESS_REGEX, value) is None:
-            raise ValueError("Invalid IP Address, must be localhost or a valid IPv4 address")
-        return value
-
-    @field_validator("sending_port", "receiver_port")
-    def port_validator(cls, value: int) -> int:
-        if value < 1 or value > 65535:
-            raise ValueError("Port must be between 1 and 65535")
-        return value
-
-
-class CameraConfig(BaseModel):
-    enabled: bool = True
-    capture_source: str = ""
-    threshold: int = 50
-    focal_length: int = 30
-    rotation_angle: int = 0
-    flip_x_axis: bool = False
-    flip_y_axis: bool = False
-    roi_x: int = 0
-    roi_y: int = 0
-    roi_w: int = 0
-    roi_h: int = 0
-
-    @field_validator("roi_x", "roi_y", "roi_w", "roi_h")
-    def roi_validator(cls, value: int) -> int:
-        if value < 0:
-            raise ValueError("ROI values must be greater than 0")
-        return value
-
-    @field_validator("capture_source")
-    def capture_source_validator(cls, value: str) -> str:
-        if not value == "":
-            if re.match(IP_ADDRESS_REGEX, value) is None:
-                raise ValueError("Invalid IP Address, must be localhost or a valid IPv4 address")
-        return value
 
 
 class EyeTrackConfig(BaseModel):
